backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.21.0
keras==3.12.1
numpy==1.26.0
pillow==10.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
